Add files via upload
The script(02_syndicated_docs_cosine_similarity) is for separating those syndicated articles. A syndicated article being a piece of content that is published on multiple websites (news outlets/publications), including the original source (so in other words, duplicate or near-duplicate articles). This was done so we don't skew the topic modeling results; otherwise, we are likely to get topics that are actually just representative of one article that has been duplicated multiple times.